pyproject.toml
[project]
name = "newspaper-web"
version = "0.1.0"
description = "Django web frontend for Newspaper RSS aggregator"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.14",
    "feedparser>=6.0",
    "requests>=2.31",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.2.2",
    "markdown>=3.10.2",
    "gunicorn>=25.0",
    "readability-lxml>=0.8",
    "geoip2>=4.8",
    "user-agents>=2.2",
    "pgvector>=0.4",
    "channels>=4.3.2",
    "daphne>=4.2.1",
    "celery[redis]>=5.4",
    "django-unfold>=0.85.0",
    "Pillow>=11.0",
    "fake-useragent>=2.2.0",
]
